replace pipes-shell with pipes-shell-2, they cannot coexist because of recipe changes
