--- shells/pipes-shell-2/node/deploy/source/serve.sh
@@ -1,2 +1,0 @@
-#!/bin/sh
-node --no-warnings ./shell.js "$@"
