+++ shells/pipes-shell/bus.js
