+++ shells/pipes-shell/dispatcher.js
