+++ shells/pipes-shell/lib/pipes-arc.js
