+++ shells/pipes-shell/lib/utils.js
