+++ shells/pipes-shell/schemas.js
@@ -22,5 +22,14 @@ export const Schemas = {
         source: 'Text',
       }
     }
+  },
+  Json: {
+    tag: 'Entity',
+    data: {
+      names: ['Json'],
+      fields: {
+        json: 'Text'
+      }
+    }
   }
 };
