+++ shells/pipes-shell/smoke.js
