+++ shells/pipes-shell/web/index.html
@@ -11,15 +11,24 @@ http://polymer.github.io/PATENTS.txt
 <html lang="en">
 <head>
 
-  <title>Pipes::Arcs</title>
+  <title>Pipes (Arcs)</title>
 
   <meta charset="utf-8">
   <meta name="viewport" content="width=device-width, user-scalable=no, minimum-scale=1.0, maximum-scale=1.0">
 
   <script src="../../configuration/csp.js"></script>
-  <!-- <link rel="stylesheet" href="../../configuration/index.css"> -->
 
   <script type="module" src="./web.js"></script>
+
+  <style>
+    body, input, button {
+      font-family: sans-serif;
+    }
+    [slotid="root"] {
+      padding: 4px;
+      border: 1px dotted gray;
+    }
+  </style>
 
 </head>
 <body>
