+++ shells/pipes-shell/web/web.js
@@ -10,54 +10,49 @@
 
 // configure
 import '../../lib/platform/loglevel-web.js';
-import {paths} from './paths.js';
-import {manifest} from './config.js';
+import {version, paths, storage} from './config.js';
 
 // optional
-import '../../lib/database/pouchdb-support.js';
-import '../../lib/database/firebase-support.js';
-//import '../../../node_modules/sourcemapped-stacktrace/dist/sourcemapped-stacktrace.js';
-
-import {Utils} from '../../lib/runtime/utils.js';
-import {ShellApiFactory} from '../device.js';
+//import '../../lib/database/pouchdb-support.js';
+//import '../../lib/database/firebase-support.js';
+import '../../configuration/whitelisted.js';
 import {DevtoolsSupport} from '../../lib/runtime/devtools-support.js';
 
-// usage:
-//
-// ShellApi.observeEntity(`{"type": "address", "name": "East Mumbleton"}`)
-// [arcid =] ShellApi.receiveEntity(`{"type": "com.google.android.apps.maps"}`)
-//
-// [arcid =] ShellApi.receiveEntity(`{"type": "com.music.spotify"}`)
-//
-// results returned via `DeviceClient.foundSuggestions(arcid, json)` (if it exists)
-
-// can be used for testing:
-//
-// window.DeviceClient = {
-//   shellReady() {
-//     console.warn('context is ready!');
-//   },
-//   foundSuggestions(arcid, json) {
-//   }
-// };
-  window.onclick = () => {
-    if (window.ShellApi) {
-      window.ShellApi.receiveEntity();
-    }
-  };
-//
-
-const storage = `pouchdb://local/arcs/`;
-const version = `version: mar-26`;
+// dependencies
+import {DomSlotComposer} from '../../lib/components/dom-slot-composer.js';
+import {RamSlotComposer} from '../../lib/components/ram-slot-composer.js';
+import {findContainers} from '../lib/utils.js';
+import {initPipe} from '../pipe.js';
+import {test} from '../smoke.js';
 
 console.log(`${version} -- ${storage}`);
+
+const composerFactory = modality => {
+  switch (modality) {
+    case 'dom': {
+      const node = document.body.appendChild(document.createElement('div'));
+      node.style = 'margin-bottom: 8px;';
+      node.innerHTML = '<div slotid="root"></div>';
+      return new DomSlotComposer({containers: findContainers(node)});
+    }
+    default:
+      return new RamSlotComposer();
+  }
+};
+
+const client = window.client || {};
 
 (async () => {
   // if remote DevTools are requested, wait for connect
   await DevtoolsSupport();
-  // configure arcs environment
-  Utils.init(paths.root, paths.map);
-  // configure ShellApi (window.DeviceClient is bound in by outer process, otherwise undefined)
-  window.ShellApi = await ShellApiFactory(storage, manifest, window.DeviceClient);
+  // configure pipes and get a bus
+  const bus = await initPipe(client, paths, storage, composerFactory);
+  // export bus
+  window.ShellApi = bus;
+  // run some examples
+  test(bus);
+  // world's dumbest ui
+  window.onclick = () => {
+    bus.receive({message: 'ingest', modality: 'dom', entity: {type: 'caption', name: 'Dogs are awesome'}});
+  };
 })();
-
